sidebar and settings screens updated
