+++ frontend/src/constants.ts
@@ -28,22 +28,7 @@ const Constants: ConstantsType = {
             path: '/logout',
             isAuth: true
         },
-        ACCOUNT: {
-            path: '/account',
-            isAuth: true
-        },
-        ACCOUNT_CHANGE_PASSWORD: {
-            path: '/account/password',
-            isAuth: true
-        },
-        SETTINGS_USER: {
-            path: '/users',
-            isAuth: true
-        },
-        SETTINGS_ADD_USER: {
-            path: '/users/add',
-            isAuth: true
-        },
+        // APP
         HOME: {
             path: '/',
             isAuth: true
@@ -74,6 +59,27 @@ const Constants: ConstantsType = {
         },
         DB_HISTORY: {
             path: '/db/[id]/history',
+            isAuth: true
+        },
+        // SETTINGS
+        SETTINGS: {
+            path: '/settings',
+            isAuth: true
+        },
+        SETTINGS_ACCOUNT: {
+            path: '/settings/account',
+            isAuth: true
+        },
+        SETTINGS_ACCOUNT_CHANGE_PASSWORD: {
+            path: '/settings/account/password',
+            isAuth: true
+        },
+        SETTINGS_USERS: {
+            path: '/settings/users',
+            isAuth: true
+        },
+        SETTINGS_ADD_USER: {
+            path: '/settings/users/add',
             isAuth: true
         },
     }
