--- frontend/src/pages/account/index.tsx
@@ -1,111 +1,0 @@
-import axios from 'axios'
-import type { NextPage } from 'next'
-import Link from 'next/link'
-import React, { useEffect, useState } from 'react'
-import AppLayout from '../../components/layouts/applayout'
-import ProfileImage, { ProfileImageSize } from '../../components/user/profileimage'
-import Constants from '../../constants'
-import { User } from '../../data/models'
-import { editUser, selectCurrentUser } from '../../redux/currentUserSlice'
-import { useAppDispatch, useAppSelector } from '../../redux/hooks'
-
-const AccountPage: NextPage = () => {
-
-  const currentUser: User = useAppSelector(selectCurrentUser)
-
-  const [editableUser, setEditableUser] = useState<User | null>(null)
-  const [saving, setSaving] = useState(false)
-  const [savingError, setSavingError] = useState(false)
-
-  const dispatch = useAppDispatch()
-
-  useEffect(()=>{
-    setEditableUser(currentUser)
-  }, [currentUser])
-
-  const startSaving = async () => {
-    setSaving(true)
-    try {
-      await dispatch(editUser({name: editableUser!.name ?? '', profileImageUrl: editableUser!.profileImageUrl})).unwrap()
-    }catch(e){
-      setSavingError(true)
-    }
-    setSaving(false)
-  }
-
-  const refreshImageUrl = async () => {
-    const response = await axios.get('https://source.unsplash.com/random/200x200')
-    const newImageURL = response.request.responseURL
-    setEditableUser({...editableUser!, profileImageUrl: newImageURL})
-  }
-
-  const revertImageUrl = async () => {
-    setEditableUser({...editableUser!, profileImageUrl: currentUser.profileImageUrl})
-  }
-
-  return (
-    <AppLayout title="Your Account | Slashbase">
-      <h1>Your Account</h1>
-      <br/>
-      { editableUser && 
-        <div className="form-container"> 
-            <div className="field">
-                <label className="label">Profile Image:</label>
-                <ProfileImage imageUrl={editableUser.profileImageUrl} size={ProfileImageSize.MEDIUM}/>
-                &nbsp;&nbsp;&nbsp;
-                <button className="button is-small" onClick={refreshImageUrl}>
-                  <span className="icon is-small">
-                    <i className="fas fa-sync"></i>
-                  </span>
-                  <span>Change</span>
-                </button>
-                &nbsp;&nbsp;
-                <button className="button is-small" onClick={revertImageUrl}>
-                  <span className="icon is-small">
-                    <i className="fas fa-undo"></i>
-                  </span>
-                  <span>Revert</span>
-                </button>
-            </div>
-            <div className="field">
-                <label className="label">Email:</label>
-                <div className="control">
-                    <input 
-                        className="input" 
-                        type="email" 
-                        value={editableUser.email}
-                        disabled={true}
-                        onChange={(e: React.ChangeEvent<HTMLInputElement>)=>{setEditableUser({...editableUser,email: e.target.value})}}
-                        placeholder="Enter your email" />
-                </div>
-            </div>
-            <div className="field">
-                <label className="label">Name:</label>
-                <div className="control">
-                    <input 
-                        className="input" 
-                        type="text" 
-                        value={editableUser.name ?? ''}
-                        onChange={(e: React.ChangeEvent<HTMLInputElement>)=>{setEditableUser({...editableUser,name: e.target.value})}}
-                        placeholder="Enter your name" />
-                </div>
-            </div>
-            {savingError && <span className="help is-danger">There was some problem saving!</span> }
-            <div className="control">
-                { !saving && <button className="button is-primary" onClick={startSaving}>Save</button> }
-                { saving && <button className="button is-primary is-loading">Saving...</button>}
-            </div>
-            <br />
-            <hr/>
-            <Link href={Constants.APP_PATHS.ACCOUNT_CHANGE_PASSWORD.path} as={Constants.APP_PATHS.ACCOUNT_CHANGE_PASSWORD.path}>
-              <a>
-                <button className="button is-seconday">Change password</button> 
-              </a>
-            </Link>
-        </div>
-      }
-    </AppLayout>
-  )
-}
-
-export default AccountPage
