+++ frontend/src/pages/settings/index.tsx
@@ -1,0 +1,15 @@
+import type { NextPage } from 'next'
+import React, { useRef, useState } from 'react'
+import AppLayout from '../../components/layouts/applayout'
+
+const SettingsPage: NextPage = () => {
+
+    return (
+        <AppLayout title="Settings | Slashbase">
+            <h1>Settings</h1>
+            <br />
+        </AppLayout>
+    )
+}
+
+export default SettingsPage
