+++ frontend/src/pages/settings/users/index.tsx
@@ -4,13 +4,13 @@ import Link from 'next/link'
 import React, { useEffect, useState } from 'react'
 import toast from 'react-hot-toast'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import UserCard from '../../components/cards/usercard/usercard'
-import AppLayout from '../../components/layouts/applayout'
-import Constants from '../../constants'
-import { User } from '../../data/models'
-import apiService from '../../network/apiService'
-import { selectCurrentUser } from '../../redux/currentUserSlice'
-import { useAppSelector } from '../../redux/hooks'
+import UserCard from '../../../components/cards/usercard/usercard'
+import AppLayout from '../../../components/layouts/applayout'
+import Constants from '../../../constants'
+import { User } from '../../../data/models'
+import apiService from '../../../network/apiService'
+import { selectCurrentUser } from '../../../redux/currentUserSlice'
+import { useAppSelector } from '../../../redux/hooks'
 
 const UsersPage: NextPage = () => {
 
@@ -19,9 +19,9 @@ const UsersPage: NextPage = () => {
   const router = useRouter()
 
   const [users, setUsers] = useState<User[]>([])
-  const [usersNext, setUsersNext] = useState<number>(0) 
+  const [usersNext, setUsersNext] = useState<number>(0)
 
-  useEffect(()=>{
+  useEffect(() => {
     if (currentUser) {
       if (currentUser.isRoot) {
         fetchUsers()
@@ -32,15 +32,15 @@ const UsersPage: NextPage = () => {
   }, [currentUser, usersNext])
 
   const fetchUsers = async () => {
-    if (usersNext === -1){
+    if (usersNext === -1) {
       return
     }
     let result = await apiService.getUsers(usersNext)
-    if(result.success) {
-      if(usersNext === 0) {
+    if (result.success) {
+      if (usersNext === 0) {
         setUsers(result.data.list)
       } else {
-      setUsers([...users,...result.data.list])
+        setUsers([...users, ...result.data.list])
       }
       setUsersNext(result.data.next)
     } else {
@@ -52,24 +52,24 @@ const UsersPage: NextPage = () => {
     <AppLayout title="Manage Users | Slashbase">
       <h1>Manage Users</h1>
       <InfiniteScroll
-          dataLength={users.length}
-          next={fetchUsers}
-          hasMore={usersNext !== -1}
-          loader={
-              <p>Loading...</p>
-          }
-          scrollableTarget="mainContainer"
-          >
-            {users.map((user: User) => (
-              <UserCard  key={user.id} user={user}/>
-            ))}
+        dataLength={users.length}
+        next={fetchUsers}
+        hasMore={usersNext !== -1}
+        loader={
+          <p>Loading...</p>
+        }
+        scrollableTarget="mainContainer"
+      >
+        {users.map((user: User) => (
+          <UserCard key={user.id} user={user} />
+        ))}
       </InfiniteScroll>
       <Link href={Constants.APP_PATHS.SETTINGS_ADD_USER.path} as={Constants.APP_PATHS.SETTINGS_ADD_USER.path}>
         <a>
           <button className="button" >
-              <i className={"fas fa-user-plus"}/>
-              &nbsp;&nbsp;
-              Add New User
+            <i className={"fas fa-user-plus"} />
+            &nbsp;&nbsp;
+            Add New User
           </button>
         </a>
       </Link>
